fix: biome warn
